Sprint 01

See merge request iseplife/api!56

+++ src/main/java/com/iseplife/api/services/GalleryService.java
@@ -1,13 +1,18 @@
 package com.iseplife.api.services;
 
+import com.iseplife.api.constants.EmbedType;
+import com.iseplife.api.constants.PostState;
 import com.iseplife.api.dao.gallery.GalleryFactory;
 import com.iseplife.api.dao.gallery.GalleryRepository;
 import com.iseplife.api.dao.media.image.ImageRepository;
+import com.iseplife.api.dao.post.PostRepository;
 import com.iseplife.api.dto.gallery.GalleryDTO;
 import com.iseplife.api.dto.gallery.view.GalleryPreview;
 import com.iseplife.api.dto.gallery.view.GalleryView;
+import com.iseplife.api.entity.Thread;
 import com.iseplife.api.entity.club.Club;
 import com.iseplife.api.entity.event.Event;
+import com.iseplife.api.entity.post.Post;
 import com.iseplife.api.entity.post.embed.media.Image;
 import com.iseplife.api.entity.post.embed.Gallery;
 import com.iseplife.api.exceptions.AuthException;
@@ -35,7 +40,13 @@ public class GalleryService {
   ImageRepository imageRepository;
 
   @Autowired
+  PostRepository postRepository;
+
+  @Autowired
   PostService postService;
+
+  @Autowired
+  StudentService studentService;
 
   @Autowired
   MediaService mediaService;
@@ -86,6 +97,10 @@ public class GalleryService {
     if (dto.getPseudo() && dto.getImages().size() > PSEUDO_GALLERY_MAX_SIZE)
       throw new IllegalArgumentException("pseudo gallery can't have more than " + PSEUDO_GALLERY_MAX_SIZE + " images");
 
+
+    gallery.setCreation(new Date());
+    gallery.setPseudo(dto.getPseudo());
+    gallery.setFeed(feedService.getFeed(dto.getFeed()));
     if (!dto.getPseudo()) {
       Club club = clubService.getClub(dto.getClub());
       if (!AuthService.hasRightOn(club))
@@ -93,21 +108,31 @@ public class GalleryService {
 
       gallery.setClub(club);
       gallery.setName(dto.getName());
+      gallery.setDescription(dto.getDescription());
     }
 
-    gallery.setCreation(new Date());
-    gallery.setPseudo(dto.getPseudo());
-    gallery.setFeed(feedService.getFeed(dto.getFeed()));
+    galleryRepository.save(gallery);
 
     Iterable<Image> images = imageRepository.findAllById(dto.getImages());
     images.forEach(img -> {
       if (img.getGallery() == null && img.getName().startsWith("img/g"))
         img.setGallery(gallery);
     });
-    gallery.setImages((List<Image>) images);
-
-    galleryRepository.save(gallery);
     imageRepository.saveAll(images);
+    if(!dto.getPseudo() && dto.getGeneratePost()){
+      Post post = new Post();
+      post.setFeed(gallery.getFeed());
+      post.setThread(new Thread());
+      post.setDescription(gallery.getDescription());
+      post.setEmbed(gallery);
+      post.setAuthor(studentService.getStudent(AuthService.getLoggedId()));
+      post.setLinkedClub(gallery.getClub());
+      post.setCreationDate(new Date());
+      post.setState(PostState.READY);
+
+      postRepository.save(post);
+    }
+
     return GalleryFactory.toView(gallery);
   }
 
@@ -133,9 +158,15 @@ public class GalleryService {
 
 
   public Boolean deleteGallery(Gallery gallery) {
+    if (!AuthService.hasRightOn(gallery))
+      throw new AuthException("You have not sufficient rights on this gallery (id:" + gallery + ")");
+
     gallery
       .getImages()
       .forEach(img -> mediaService.deleteMedia(img));
+    gallery.setImages(null);
+    List<Post> p = postRepository.findAllByEmbed(gallery.getId(), EmbedType.GALLERY);
+    postRepository.deleteAll(p);
 
     galleryRepository.delete(gallery);
     return true;
@@ -151,6 +182,8 @@ public class GalleryService {
       .stream()
       .filter(img -> images.contains(img.getId()))
       .count();
+
+
 
     if (images.size() != imageSize) {
       throw new IllegalArgumentException("images does not belong to this gallery");
